// src/app/services/models/employee-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DepartementDto } from '../models/departement-dto';
import { EmployeeShortDto } from '../models/employee-short-dto';
import { PositionDto } from '../models/position-dto';
export interface EmployeeDto {
  address?: string;
  bankAccountInfo?: string;
  birthDate?: string;
  city?: string;
  cv?: string;
  departement?: DepartementDto;
  emergencyContactName?: string;
  emergencyContactPhone?: string;
  endDate?: string;
  firstName?: string;
  gender?: 'MALE' | 'FEMALE';
  hireDate?: string;
  id?: number;
  lastName?: string;
  manager?: EmployeeShortDto;
  phone?: string;
  position?: PositionDto;
  socialSecurityNumber?: string;
  status?: 'ACTIVE' | 'INACTIVE' | 'ON_LEAVE' | 'TERMINATED';
}