// src/app/services/models/document-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface DocumentDto {
  contentType?: string;
  downloadUrl?: string;
  id?: number;
  name?: string;
  path?: string;
  size?: number;
  type?: 'CV' | 'IDENTITY' | 'DIPLOMA' | 'CONTRACT' | 'PAYSLIP' | 'CERTIFICATE' | 'ADMINISTRATIVE' | 'OTHER';
  uploadDate?: string;
}